+++ src/cli.php
@@ -12,6 +12,7 @@ try {
         'new' => (include __DIR__ . '/cli-new.php')(config($cpath)),
         'list' => (include __DIR__ . '/cli-list.php')(config($cpath)),
         'edit' => (include __DIR__ . '/cli-edit.php')(config($cpath), array_slice($argv, 2)),
+        'delete' => (include __DIR__ . '/cli-delete.php')(config($cpath), array_slice($argv, 2)),
         'enable' => (include __DIR__ . '/cli-onoff.php')(config($cpath), array_slice($argv, 2), true),
         'disable' => (include __DIR__ . '/cli-onoff.php')(config($cpath), array_slice($argv, 2), false),
         'test' => (include __DIR__ . '/cli-test.php')(config($cpath), array_slice($argv, 2)),